refactor(auth): extract tenant login
